[feat] 질문 생성 시 포인트 지급 기능 구현

+++ rod/src/main/java/com/example/rod/answer/dto/AnswerRequestDto.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class AnswerRequestDto {
     private String content;
 
-    private float difficulty;
+    private double difficulty;
 
 }
